[Change] separate admin controller functionality from router
