+++ index.js
@@ -99,8 +99,8 @@ app.get("*", checkDB, (req, res) => {
 app.use((err, req, res, next) => {
     const { status = 500, message = "Something went wrong" } = err;
 
-    res.status(status).send(err);
     logger.error(`status: ${status}; message: ${message}`);
+    res.status(status).send(err);
 });
 
 app.listen(port, () => {
